% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_summary.R
\name{report_summary}
\alias{report_summary}
\title{Calculating mean values for each virus and each day}
\usage{
report_summary(growth)
}
\arguments{
\item{growth}{\link{\code{character(1)}}\cr Excel-file with 3 replicates}
}
\value{
character string hat gives mean values for each day and each virus
}
\description{
Calculating mean values for each virus and each day
}
